requirements.txt
black==22.3.0
dj-database-url==0.5.0
django==4.0.3
django-environ==0.9.0
django-extensions==3.1.5
django-on-heroku==1.1.2
flake8==4.0.1
gunicorn==20.1.0
icecream==2.1.2
psycopg2-binary==2.9.3
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-selenium==3.0.0
python-decouple==3.6
python-dotenv==0.20.0
requests==2.27.1
selenium==3.141.0
sentry-sdk==1.8.0
whitenoise==6.1.0
